Massive Change List as follows:
Code Optimization:
Made a ProtectedBlocks a global static for the class, so that it's not initialized each event, considering it's the most used event.
Made parameters on all methods final.
Made any variable capable of being final as such.
Removed the WoodenDoor randomization, as it was already the default.

Code Style:
Mostly whitespace and spacing changes.
Moved if statements without curly braces into the lines following for easier readability.

Additional Changes:
Removed the unused level variable for bit drops. The comment states that d20 is checked due to blockchain size, but that variable was never actually used.
Commented out the onAttack event's code, as it did nothing.