Massive Commit: Recreated commit branch after merge issues.
replace and replaceAll in Java return modified string, as Strings are immutable. Therefore, update the object.
Moved scheduler to after the null check.
Spacing between various ==, etc..
Made ProtectedBlocks a global static for the class, so that it's not initialized each event, considering it's the most used event.
Made parameters on all methods final.
Made any variable capable of being final as such.
Removed the WoodenDoor randomization, as it was already the default.
Mostly whitespace and spacing changes.
Moved if statements without curly braces into the lines following for easier readability.
Removed the unused level variable for bit drops. The comment states that d20 is checked due to blockchain size, but that variable was never actually used.
Commented out the onAttack event's code, as it did nothing.
Attempted to address issue #47 with your idea of sending "the wilderness" as a gray message to distinguish it.