Added very demanded items to the Market Inventory.
Since the nether is blocked, Netherwart is very demanded in the server specially by pro players who farm bits the hardest. I propose to sell netherwart seeds, but without a very expensive price tag, so they also buy soul sand(already in the inventory) which is required to grow the netherwart.
Netherwart can't be crafted or taken from the nether, so 400 bits for 16 netherwart seeds looks very nice. Feel free to adjust it, 
400 + 200 from the soul sand = 600 bits to the loot wallet.

Also, so many people tell me that they are willing to pay anything to get notch apples because those cannot be crafted anymore. I propose to sell notch apples. 6 notch apples for 600 bits I think is fair enough: only 2 times the price of the normal golden apple and it recovers the loot wallet from 3 mob drops.

More:
-Added slime balls. This is probably the item that most people is currently interested in.
-Added paper. People that don't want to plant sugar canes inside their properties will buy this to make books and more.
-Added anvils.
-Added enderpearls.


Some issues: I've  found that notch apples in Bukkit are added like this-> Material.GOLDEN_APPLE, 1, (short)1; 
but I am not sure if the BitQuest plugin has tweaked that out. Please somebody check that out and make sure it is correct. We need notch apples in BitQuest! :)